refactor(LeaderboardRow): relocate component
