+++ init.js
@@ -1,5 +1,5 @@
-const fs = require('fs')
+const fs = require('fs');
 
-fs.copyFileSync('.env.sample', '.env')
+fs.copyFileSync('.env.sample', '.env');
 
-fs.mkdirSync('tmp/pgdata', { recursive: true })
+fs.mkdirSync('tmp/pgdata', { recursive: true });
